Update README.md
Starts with the source node choose the neighbour with the highest value, Move to the node with the highest value, choose the node with the highest value, and move to the next highest node value and so on. The Hill climbing starts with the source node and navigates the graph to find the destination with the highest cumulative value.